Revert "KFSPTS-13687: Move CuAward fields to extended attribute (#676)"
This reverts commit 20947c5db4a580bbdb615c28f5277fa299fa6051.
